eval/pyproject.toml
[project]
name = "reader-eval"
version = "0.1.0"
description = "Evaluation pipeline for reader LLM enrichment"
requires-python = ">=3.12"
dependencies = [
    "openai>=1.0.0",
    "google-genai>=0.2.0",
    "jsonschema>=4.0.0",
    "pyyaml>=6.0",
    "pydantic>=2.0.0",
    "tenacity>=8.0.0",
]

